requirements.txt
flask==3.0.3
line-bot-sdk==3.17.1
openai==1.93.2
python-dotenv==1.0.1
gunicorn==21.2.0
